backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
firebase-admin==6.4.0
cloudinary==1.37.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
Pillow==10.2.0
ultralytics==8.1.2
numpy==1.26.4
opencv-python==4.9.0.80